requirements.txt
flask
firebase-admin
python-dotenv
reportlab
gunicorn
flask-cors
Werkzeug
anthropic
